worker/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
pybloom-live>=3.0.0
pymongo>=4.6.0
kafka-python>=2.0.2
